+++ CHANGELOG.md
@@ -6,7 +6,7 @@ principles.
 
 ## Unreleased
 
--   Changed: Movee duplicate action from index to create page to improve index pageload
+-   Changed: Move duplicate action per index item to the create action on the index. This will improve index pageload
 -   Changed: TimeTable FieldPresets::timetable now returns a Form component instead of an iterable
 -   Removed: State options from index page (such as put online / offline / archive) to improve index pageload